Remove custom miim traits in favour of mdio crate. Fix field bit order.
Removes the custom `miim::{Read, Write}` traits in favour of using the
`mdio` traits. This allows for greater flexibility, and allows for
taking advavntage of the `mdio` crate's bitbang feature. This feature is
necessary when using the `ksz8863` with an `stm32f1` board, as these
boards don't allow for specifying a custom Op code in their MAC's MIIM
implementation.

Fixes an issue where the order of the fields were back to front due to
using `Msb0` instead of `Lsb0`. Updates tests accordingly.